+++ src/views/radios.rs
@@ -287,6 +287,7 @@ impl RadiosPage {
                 let style = SlotListSlotStyle::for_slot(
                     ctx.is_center,
                     false,
+                    false,
                     ctx.is_selected,
                     ctx.has_multi_selection,
                     ctx.opacity,