+++ src/views/songs/view.rs
@@ -186,6 +186,7 @@ impl SongsPage {
                 let style = SlotListSlotStyle::for_slot(
                     ctx.is_center,
                     false,
+                    false,
                     ctx.is_selected,
                     ctx.has_multi_selection,
                     ctx.opacity,